telegram-ingestion/requirements.txt
sqlalchemy>=2.0.25,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0
python-dotenv>=1.0.0,<2.0.0
telethon>=1.36.0,<2.0.0
cryptography>=42.0.0,<44.0.0
pypdf>=4.0.0,<6.0.0
python-docx>=1.1.0,<2.0.0
